services/ingest-api/requirements.txt
# =============================================================================
# Ingest API Service - Dependencies
# =============================================================================

# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Google Cloud
google-cloud-pubsub==2.27.1

# Data Validation
pydantic==2.10.2
pydantic-settings==2.6.1

# Utilities
python-dotenv==1.0.1
structlog==24.4.0

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
ruff==0.8.2
mypy==1.13.0
